Press releases.
Added the feeds and job_scheduler contrib modules.
Added a Feature for the feed content type.
Added a Feature for the press release content type.
Added a module providing a custom Feeds fetcher.
Switched on the rbkc_feeds module.
Added a press release template file, will need more work!
Added a View for press releases.
Added a teaser template for press releases.
Added the insert into feeds source, without this the node_save() on its
own was not enough to get the feed URL in.